Move segment and track machines into tools folder
