+++ visualizer/src/service/tools/segment/brushMachine.js
