+++ visualizer/src/service/tools/segment/thresholdMachine.js
