+++ visualizer/src/service/tools/segment/toolUtils.js
