+++ visualizer/src/service/tools/segment/trimMachine.js
