+++ visualizer/src/service/tools/trackingMachine.js
